refactor(page): Pull out 'all' paginator
