+++ src/pagination/mod.rs
@@ -12,6 +12,7 @@ use crate::document;
 use crate::document::Document;
 use crate::error::Result;
 
+mod all;
 mod categories;
 mod dates;
 mod helpers;
@@ -33,7 +34,7 @@ pub(crate) fn generate_paginators(
     match config.include {
         Include::All => {
             sort_posts(&mut all_posts, config);
-            create_all_paginators(&all_posts, doc, config, None)
+            all::create_all_paginators(&all_posts, doc, config, None)
         }
         Include::Tags => tags::create_tags_paginators(&all_posts, doc, config),
         Include::Categories => categories::create_categories_paginators(&all_posts, doc, config),
@@ -42,34 +43,6 @@ pub(crate) fn generate_paginators(
             unreachable!("PaginationConfigBuilder should have lead to a None for pagination.")
         }
     }
-}
-
-fn create_all_paginators(
-    all_posts: &[&liquid::model::Value],
-    doc: &Document,
-    pagination_cfg: &PaginationConfig,
-    index_title: Option<&liquid::model::Value>,
-) -> Result<Vec<Paginator>> {
-    let total_pages = all_posts.len();
-    // f32 used here in order to not lose information to ceil the result,
-    // otherwise we can lose an index
-    let total_indexes = (total_pages as f32 / pagination_cfg.per_page as f32).ceil() as usize;
-    let paginators: Result<Vec<_>> = all_posts
-        .chunks(pagination_cfg.per_page as usize)
-        .enumerate()
-        .map(|(i, chunk)| {
-            paginator::create_paginator(
-                i,
-                total_indexes,
-                total_pages,
-                pagination_cfg,
-                doc,
-                chunk,
-                index_title,
-            )
-        })
-        .collect();
-    paginators
 }
 
 // sort posts by multiple criteria
